7_LLM/Text_To_SQL/pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.6.0",
    "pg8000>=1.31.5",
    "python-dotenv>=1.1.1",
]
